pyproject.toml
[project]
name = "py-dashboard-dash"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
  "dash>=3,<4",
  "plotly>=5",
  "flask>=3,<4",
  "flask-caching>=2.3",
  "pandas>=1.5",
  "numpy>=1.23",
  "requests>=2.28",
  "SQLAlchemy>=1.4",
  "pyjwt>=2.4",
  "gunicorn>=21.0",
  "uvicorn-worker>=0.2",
  "pandas-stubs==2.3.2.250926",
  "python-dotenv>=1.0",
  "pydantic>=2,<3",
  "pydantic-settings>=2,<3",
]
